+++ android/app/src/main/java/com/kibalivewire/quire/MainActivity.java
@@ -1,0 +1,5 @@
+package com.kibalivewire.quire;
+
+import com.getcapacitor.BridgeActivity;
+
+public class MainActivity extends BridgeActivity {}